finished work on api automodule for docs
