+++ docs/api/interfaces.behavior.rst
@@ -1,0 +1,7 @@
+Behavior
+==========
+
+Movie Data
+---------------
+
+.. automodule:: nwb_conversion_tools.datainterfaces.behavior.movie.moviedatainterface
