+++ nwb_conversion_tools/datainterfaces/ecephys/tutorial/recordingtutorialdatainterface.py
@@ -1,15 +1,11 @@
 """Authors: Cody Baker and Ben Dichter."""
-from pathlib import Path
-
-import numpy as np
 import spikeextractors as se
 
 from ....utils.json_schema import get_schema_from_method_signature
 from ..baserecordingextractorinterface import BaseRecordingExtractorInterface
-from ..basesortingextractorinterface import BaseSortingExtractorInterface
 
 
-class TutorialRecordingInterface(BaseRecordingExtractorInterface):
+class RecordingTutorialInterface(BaseRecordingExtractorInterface):
     """High-pass recording data interface for demonstrating NWB Conversion Tools usage in tutorials."""
 
     RX = se.NumpyRecordingExtractor
@@ -26,7 +22,6 @@ class TutorialRecordingInterface(BaseRecordingExtractorInterface):
         self.source_data = source_data
 
     def get_metadata(self):
-        """Auto-populate extracellular electrophysiology metadata."""
         metadata = dict(
             Ecephys=dict(
                 Device=[
@@ -57,34 +52,5 @@ class TutorialRecordingInterface(BaseRecordingExtractorInterface):
                     description="Raw acquisition traces for the NWB Conversion Tools tutorial."
                 )
             )
-        )
-        return metadata
-
-
-class TutorialSortingInterface(BaseSortingExtractorInterface):
-    """Sorting data interface for demonstrating NWB Conversion Tools usage in tutorials."""
-
-    SX = se.NumpySortingExtractor
-
-    @classmethod
-    def get_source_schema(cls):
-        source_schema = get_schema_from_method_signature(se.example_datasets.toy_example)
-        source_schema['additionalProperties'] = True
-        return source_schema
-
-    def __init__(self, **source_data):
-        self.sorting_extractor = se.example_datasets.toy_example(**source_data)[1]
-        self.source_data = source_data
-
-    def get_metadata(self):
-        """Auto-populate unit table metadata."""
-        metadata = dict(
-            UnitProperties=[
-                dict(
-                    name="custom_unit_column",
-                    description="Custom column in the spiking unit table for the NWB Conversion Tools tutorial.",
-                    data=[x for x in range(len(self.sorting_extractor.get_unit_ids()))]
-                )
-            ]
         )
         return metadata
